[~] geometry: Use GLuint instead of GLushort for IndexBuffer

+++ src/core/include/ttb/core/geometry/IndexBuffer.hpp
@@ -14,7 +14,7 @@ namespace ttb
     class IndexBuffer
     {
     public:
-        using Index = GLushort;
+        using Index = GLuint;
 
         static std::shared_ptr< IndexBuffer > create();
